#410: added duration to latest commit for a remote branch to mob status
